requirements.txt
aiohttp==3.4.2
bleak==0.10.0
Flask==1.1.2
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
Jinja2==2.11.3
pycparser==2.20
SQLAlchemy==1.3.23
tzlocal==2.1
update==0.0.1
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
yarg==0.1.9
python-dotenv